[refactor] Rename to KLLVM
