+++ pom.xml
@@ -4,17 +4,17 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>dev.supergrecko.ktools</groupId>
-    <artifactId>parser</artifactId>
+    <groupId>dev.supergrecko</groupId>
+    <artifactId>kllvm</artifactId>
     <version>0.0.0</version>
 
     <properties>
-        <exec.mainClass>dev.supergrecko.ktools.llvm.MainKt</exec.mainClass>
+        <exec.mainClass>dev.supergrecko.kllvm.MainKt</exec.mainClass>
 
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
 
-        <project.repository>https://github.com/kompiler-tools/parser</project.repository>
+        <project.repository>https://github.com/kompiler-tools/kllvm</project.repository>
 
         <jvm.target>1.8</jvm.target>
         <kotlin.version>1.3.70</kotlin.version>
@@ -111,7 +111,7 @@
                         <configuration>
                             <archive>
                                 <manifest>
-                                    <mainClass>dev.supergrecko.ktools.parser.MainKt</mainClass>
+                                    <mainClass>dev.supergrecko.kllvm.MainKt</mainClass>
                                 </manifest>
                             </archive>
                             <descriptorRefs>
