+++ src/main/kotlin/dev/supergrecko/kllvm/Main.kt
@@ -1,0 +1,5 @@
+package dev.supergrecko.kllvm
+
+fun main() {
+    println("Hello World!")
+}
