--- src/main/kotlin/dev/supergrecko/ktools/llvm/Main.kt
@@ -1,3 +1,0 @@
-package dev.supergrecko.ktools.llvm
-
-fun main() = println("Hello World")
